+++ inst/htmlwidgets/lib/circles/circles.js
@@ -16,7 +16,7 @@ function add_circles(map_id, data_circles, update_map_view, layer_id, use_polyli
     if (focus === true) {
         clear_bounds(map_id);
     }
-    
+
     var i,
         infoWindow = new google.maps.InfoWindow();
 
@@ -102,6 +102,8 @@ function set_circle(map_id, circle, infoWindow, update_map_view, layer_id, use_p
 }
 
 function set_each_circle(Circle, circle, infoWindow, update_map_view, map_id, layer_id) {
+
+  console.log(Circle);
 
     var shapeInfo;