+++ inst/htmlwidgets/lib/kml/kml.js
@@ -1,14 +1,14 @@
 function add_kml(map_id, kml_data, layer_id, update_map_view) {
 
-  //window[map_id + 'googleKml' + layer_id] = [];
-  createWindowObject(map_id, 'googleKml', layer_id);
+    //window[map_id + 'googleKml' + layer_id] = [];
+    createWindowObject(map_id, 'googleKml', layer_id);
 
-  console.log(update_map_view);
-  var kmlLayer = new google.maps.KmlLayer({
-    url: kml_data[0].url,
-    preserveViewport: update_map_view,
-    map: window[map_id + 'map']
-  });
+    var kmlLayer = new google.maps.KmlLayer({
+        url: kml_data[0].url,
+        map: window[map_id + 'map'],
+        preserveViewport: update_map_view,
+        zIndex: kml_data[0].z_index
+    });
 
   window[map_id + 'googleKml' + layer_id].push(kmlLayer);
 }